Moved content from plotting table scores back to annotate.py
